Update #2
Fixed a few instances where it would still try to fly. 

Optimized the random node checking a bit more.

Should teleport home less when starting but there are still places in the base (pastures, farms, next to workshop) that it can't navigate from/to so it will cast a return then.

Current Order is now saved between RB restarts.

There's still that occasion 30 second hang when it searches for random nodes with the new system. I've asked Mastahg about it. I still think the new system is more efficient if you're gathering anything that is in an area that involve vertical height differences.